Improves console output

+++ src/main/java/p/lodz/dashboardsimulator/utils/Utils.java
@@ -5,6 +5,6 @@ public class Utils {
     public static double round(double value, int n) {
         double places = Math.pow(10.0, n);
 
-        return ((double) Math.round(value * places)) / places;
+        return Math.round(value * places) / places;
     }
 }
